fix: data classes

+++ app/src/main/java/com/example/readerapp/models/VolumeInfo.kt
@@ -1,9 +1,11 @@
 package com.example.readerapp.models
 
+import androidx.collection.FloatSet
+
 data class VolumeInfo(
     val allowAnonLogging: Boolean,
     val authors: List<String>,
-    val averageRating: Int,
+    val averageRating: Double,
     val canonicalVolumeLink: String,
     val categories: List<String>,
     val contentVersion: String,
